Split modal to overlay and frame
[MAILPOET-2777]
